requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asyncpg==0.30.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
Django==5.1.6
django-ninja==1.3.0
dnspython==2.7.0
email_validator==2.2.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
load-dotenv==0.1.0
packaging==24.2
pluggy==1.5.0
psycopg==3.2.4
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.4
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
